+++ packages/core/src/components/dropdown-menu/dropdown-menu.tsx
@@ -64,9 +64,13 @@ export class DropdownMenu {
     options.setAttribute("aria-labelledby", this.button.id);
 
     for (const ul of Array.from(this.host.getElementsByTagName("ul"))) {
-      ul.setAttribute("role", "none");
+      ul.setAttribute("role", "group");
       for (const li of Array.from(ul.getElementsByTagName("li"))) {
-        li.setAttribute("role", "none");
+        if (li.classList.contains("dso-group-label")) {
+          li.setAttribute("role", "menuitem");
+        } else {
+          li.setAttribute("role", "none");
+        }
       }
     }
   }
@@ -78,7 +82,7 @@ export class DropdownMenu {
         if (this.checkable && li.classList.contains("dso-checked")) {
           tab.setAttribute("aria-checked", "true");
         } else {
-          tab.removeAttribute("aria-checked");
+          tab.setAttribute("aria-checked", "false");
         }
       }
     }