+++ packages/dso-toolkit/src/components/dropdown-menu/dropdown-menu.models.ts
@@ -9,6 +9,7 @@ export interface DropdownMenu {
 }
 
 export interface DropdownMenuGroup {
+  id?: string;
   header?: string;
   items: Array<DropdownMenuItem & (DropdownMenuItemAnchor | DropdownMenuItemButton)>;
 }